Split CityPage into logical and presentational components
